requirements.txt
django==2.2.2
djangorestframework==3.9.4
django-filter==2.1.0
mysqlclient==1.4.2